requirements.txt
ultralytics
opencv-python
Pillow
torch